Feat : 캐릭터 관련 함수 클래스 생성
